requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pandas>=2.2.0,<3.0.0
numpy>=1.26.0,<3.0.0
scikit-learn>=1.5.0,<2.0.0
openai>=1.40.0,<2.0.0
joblib>=1.4.0,<2.0.0
pydantic-settings>=2.4.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pytest>=8.0.0,<9.0.0
httpx>=0.27.0,<1.0.0
